+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project are documented here. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Optional light-gray minimap dots for dungeon monsters that are not selected in any pull of the route. The color and opacity can be customized in Pull Colors.
+
 ## [1.4.5] - 2026-08-12
 
 ### Fixed